13/03/18 corrected parent selection select_parents
Modified select_parents() so that it compares parent_selector
to check if it is less than the point in accumulated_reciprocals.

Previously it looked to see if it was larger and I think this was
wrong - over all not an issue as due to the way the parent or daughter is
selected.

Also changed a few:
for i in range(len(a)):
    something a[i]